Split SerializeTest into multiple files
