+++ fluent.syntax/src/test/kotlin/org/projectfluent/syntax/serializer/SerializeEntryTest.kt
@@ -1,0 +1,17 @@
+package org.projectfluent.syntax.serializer
+
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+
+class SerializeEntryTest : SerializerTest() {
+
+    @Test
+    fun message() {
+        val input = """
+            # Attached comment
+            key = Value
+            
+        """.trimIndent()
+        assertEquals(input, this.pretty(input))
+    }
+}
